### 2019-04-14 #1378 @labarba
👋 @prabhuramachandran — have a look at the tabs on the metadata section of your paper. Looks like some weird characters are messing up the paper compilation.
