### 2019-04-14 #1378 @labarba
@jasonclark — this submission looks to be in your areas of expertise and interest. I'm hoping you can handle it. I'll assign you now, but do let me know if you're disinclined for any reason.
