### 2019-05-29 #1378 @ethanwhite
Cool idea - unfortunately I'm not available to review in the near future.
