### 2019-06-03 #1378 @kyleniemeyer
@deniederhut great! We don't have a specific deadline for reviews, but generally we like to get reviews within two or three weeks—but longer can be ok.
