### 2019-06-03 #1378 @deniederhut
Okay, good to know. We're in the middle of the review period for SciPy, so it be a few weeks before I can take a look.
