### 2019-06-12 #1501 @whedon
[REVIEW]: View eXtract aNnotate (ViXeN) media files: a multimedia project manager
**Submitting author:** @prabhuramachandran (<a href="http://orcid.org/0000-0001-6337-1720">Prabhu Ramachandran</a>)
**Repository:** <a href="https://github.com/vixen-project/vixen" target ="_blank">https://github.com/vixen-project/vixen</a>
**Version:** 1.0rc3

- [x] **Version:** 1.0rc3

- [x] **Version:** 1.0rc3
