### 2019-06-12 #1378 @arfon
@deniederhut, @amoeba - please head over to #1501 where the review will take place.
