### 2019-06-17 prabhuramachandran/vixen#23 @amoeba
JOSS review
Hi @prabhuramachandran, I've just finished going through the review checklist over on https://github.com/openjournals/joss-reviews/issues/1501 and only had one item I couldn't check off:

Could you please expand the "Support" section of the README to cover this?

Otherwise, everything else looks great. I found a few typographical errors that I'll send a PR for in a sec. Let me know if you have any questions and also let me know when you want me to look at the README again.

I found the software pretty straightforward to use but had one question: In the Project view, I tried to select multiple items and couldn't. Is there a way or could this be a useful thing to add? This question isn't strictly part of the review process so you don't need to answer it or do anything about it in order to finalize the review.
