### 2019-06-17 1826420
Fix some minor typos
Found as part of my review process,

https://github.com/openjournals/joss-reviews/issues/1501

and contributed for the authors.
