### 2019-06-17 #1501 @amoeba
I'm almost done and had one review checklist item I couldn't check off: https://github.com/vixen-project/vixen/issues/23. This review is a Conditional Accept for now.

@prabhuramachandran, can you please take a look at https://github.com/vixen-project/vixen/issues/23 and let me know what you think? Overall, ViXeN looks and works great. You've clearly put a lot of time and effort into building it and testing it. I found the installation as easy as you outlined, running it to be straight-forward, and creating a test project for tagging some screenshots I had laying around was easy (esp. with your docs). The [documentation](https://vixen.readthedocs.io/en/latest/) was well-written and was useful as I went. The ability to add tags of various data types and add Processors is really nice. I can see myself recommending this to colleagues. As a bonus, I was happy to see what looked like a decent test suite covering the various components of the application. I sent in a PR for some minor typos but otherwise only had one issue which I've described in https://github.com/vixen-project/vixen/issues/23. Please take a look!
