### 2019-06-17 prabhuramachandran/vixen#23 @prabhuramachandran
Hi Bryce,

Thanks for the kind comments and your thorough review.  I will make the requested changes to the support section later today.  I have merged your kind PR!

Multiple selection is a great idea and one we thought about but something I am not quite sure how best to implement.  Is this just to be able to edit multiple tags?  The issue is, what exactly should the UI do on a multi-selection of files, what about a directory?  If there are already existing fields, multi-selecting them would allow a user to accidentally edit tags when they may not have wanted to.  Also it isn't clear what the video view should do, showing multiple videos/pictures may be a bit too much?  Given the difficulties in what exactly would be ideal I just decided not to implement it until there was a clear idea of what exactly we wanted to do.  I hope that clarifies the reason why we don't have that feature (yet).

Anyway, thanks for the pointer to the support section needing work, I will add that later today.

Thanks again for taking the time to review ViXeN!
