### 2019-06-17 prabhuramachandran/vixen#23 @amoeba
Sounds like you've definitely scratched your head a bit on the multiple-selection idea and I agree it isn't straightforward. The use case I was thinking of would be to tag a set of files with the same tag and value, maybe just two files that are adjacent in the same folder or maybe an entire sub-folder of files. For example, if I was tagging site codes on a set of camera trap images and I had organized my images in subfolders by site, I could imagine wanting to tag a set of files all at once.
