### 2019-06-18 359d7e6
Expanding the support section as requested in #23.
