### 2019-06-18 ea2ef77
Update issue templates
