+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,39 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**Software Version**
+Version of ViXeN and how it was installed.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Operating System (please complete the following information):**
+ - OS: [e.g. iOS]
+ - Browser [e.g. chrome, safari]
+ - Version [e.g. 22]
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Log file**
+Optionally send us your vixen log file if you are able.  See the [instructions for obtaining the log file](https://vixen.readthedocs.io/en/latest/installation.html#troubleshooting).
+
+**Additional context**
+Add any other context about the problem here.
