### 2019-06-18 prabhuramachandran/vixen#23 @prabhuramachandran
Hi @amoeba,

Thanks again for a specific use case.  This is a very useful idea indeed, we will keep this in mind for a subsequent release.  I have also expanded the support section a bit in #25.  In addition, I also just pushed a bug report template to make it easier for users to report problems and provide the necessary information.  I hope this is enough, please let me know if #25 is good enough and I can merge it, otherwise I will add to that PR with any suggestions you may have.

Thanks again!
