### 2019-07-10 #1501 @deniederhut
Oh how thoughtful! Yeah, I should have time next week.
