### 2019-07-17 #1501 @deniederhut
Hey @xuanxu ! I have a COI question. In the last year, Prabhu and I worked for the same company (Enthought). We were not working on the same project, and at no time was one of us a report of the other. Can I still review this submission?
