### 2019-07-17 #1501 @xuanxu
@deniederhut Thanks for noting it! If you are not currently affiliated with the same institution/company and you didn’t work on the same project then by the current [policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) I see at most a minor COI, that can be waived if you think you are able to make an impartial assessment of this submission.
