### 2019-07-18 #1501 @deniederhut
Okay cool. I like Prabhu a lot, but I think I can still be sufficiently impartial.

I see that I am supposed to leave comments about the codebase on its own issue tracker - where do I put comments about the paper?
