### 2019-07-18 #1501 @xuanxu
@deniederhut This is the main issue where the review takes place. You can add comments about the paper here. If you also create issues or pull requests in the submitted software’s issue tracker, please mention/link them here.
