### 2019-07-19 prabhuramachandran/vixen#27 @deniederhut
Program fails to initialize - syntax error in 3.7
**Describe the bug**
Program fails with syntax error in python >= 3.7. Depending on your philosophical perspective, this is either a bug about using a keyword as a name or a bug about allowing installs for unsupported python versions.

**Software Version**
ViXeN version: 1.0rc3.dev0, from master, via pip

**To Reproduce**
Run vixen console entry point

**Expected behavior**
Program executes

**Log file**

Relevant logs

```
INFO|2019-07-18 21:59:05,564|vixen.cli|**** Starting ViXeN ****
INFO|2019-07-18 21:59:05,564|vixen.cli|Project root: /Users/dillon/.vixen
INFO|2019-07-18 21:59:05,564|vixen.cli|ViXeN version: 1.0rc3.dev0
INFO|2019-07-18 21:59:05,564|vixen.cli|Runing from /Users/dillon/githubpackages/vixen/vixen
INFO|2019-07-18 21:59:05,564|vixen.cli|uname_result(system='Darwin', node='Majora.local', release='17.7.0', version='Darwin Kernel Version 17.7.0: Thu Jun 21 22:53:14 PDT 2018; root:xnu-4570.71.2~1/RELEASE_X86_64', machine='x86_64', processor='i386')
INFO|2019-07-18 21:59:05,565|vixen.cli|vixen called as: ['/Users/dillon/anaconda/envs/vixen-test/bin/vixen']
INFO|2019-07-18 21:59:05,565|vixen.cli|Parsed command line args: Namespace(console=False, dev=False, port=None, version=False)
ERROR|2019-07-18 21:59:05,567|vixen.cli|Uncaught exception
Traceback (most recent call last):
  File "/Users/dillon/anaconda/envs/vixen-test/bin/vixen", line 11, in <module>
    load_entry_point('vixen', 'console_scripts', 'vixen')()
  File "/Users/dillon/githubpackages/vixen/vixen/cli.py", line 95, in main
    view(dev=opts.dev, port=opts.port)
  File "/Users/dillon/githubpackages/vixen/vixen/cli.py", line 63, in view
    from vixen.vixen_ui import main
  File "/Users/dillon/githubpackages/vixen/vixen/vixen_ui.py", line 51
    async = False
          ^
SyntaxError: invalid syntax
```

**Additional context**
You may wish to consider specifying a `python_requires` in your setup.py.

xref https://github.com/openjournals/joss-reviews/issues/1501
