### 2019-07-19 prabhuramachandran/vixen#28 @deniederhut
Paper: link to the documentation
In both the readme and the paper submitted to JOSS, the documents are linked via the display name `https://vixen.rtfd.io`. This is a fun little inside joke, but the `f` stands for a word which is not generally considered professional. I might suggest using the equivalent `vixen.readthedocs.io` instead.

xref https://github.com/openjournals/joss-reviews/issues/1501
