### 2019-07-19 prabhuramachandran/vixen#29 @deniederhut
File chooser freezes UI
**Describe the bug**
In --dev mode (possibly also others) vixen freezes under Edit Project -> Path -> Browse if you cancel before choosing a filename.

**Software Version**
ViXeN version: 1.0rc3.dev0 from master via pip

**To Reproduce**

1. Start vixen in dev
2. New project
3. Browse to path
4. Click cancel

**Expected behavior**
Filechooser should close.

**Operating System (please complete the following information):**
MacOS

**Log file**
Nothing in logs about this

**Additional context**
xref https://github.com/openjournals/joss-reviews/issues/1501
