### 2019-07-19 prabhuramachandran/vixen#29 @deniederhut
Hm. It looks like actually choosing a folder might also be freezing?
