### 2019-07-19 prabhuramachandran/vixen#29 @prabhuramachandran
Thanks for the issues, I will take a look over this weekend and fix them as possible.  The issue with the web-UI is that there isn't an easy way to ask for a file on the browser so I resort to using tkinter for this and that seems to be flaky.
