### 2019-07-19 prabhuramachandran/vixen#28 @prabhuramachandran
Thanks.  I always thought the 'f' stood for fine.  ;-) I think it is unfortunate that in general offensive language is assumed when a nicer word is possible.  Having said that I will change the link to the longer version in the interest of not causing offense.  Will update the ticket when this is done.
