### 2019-07-22 0fc2b2c
Use readthedocs.io for link to documentation.
Fixes #28.
