### 2019-07-30 prabhuramachandran/vixen#27 @prabhuramachandran
I had started some work last week but I had to travel to the UK for a workshop and then fell very sick on the way back.  I have recovered but my semester just started, so I will be able to get to the issues by the coming weekend.  In the meanwhile it may be best to use the binaries to test the software perhaps?
