### 2019-08-07 187355c
This seems to fix #29.
On a Linux machine with tk version 8.6 and Tcl version 8.6 this prevents
strange freezes.
