### 2019-08-07 2f9783a
Update travis-ci to work with Xenial.
