+++ .travis.yml
@@ -1,4 +1,7 @@
 language: python
+dist: xenial
+services:
+  - xvfb
 python:
   - 2.7
   - 3.6
@@ -17,9 +20,6 @@ before_script:
   - mkdir geckodriver
   - tar -xzf geckodriver*.tar.gz -C geckodriver
   - export PATH=$PATH:$PWD/geckodriver
-  - "export DISPLAY=:99.0"
-  - "sh -e /etc/init.d/xvfb start"
-  - sleep 3 # give xvfb some time to start
 
 script:
   - coverage erase
