### 2019-08-07 prabhuramachandran/vixen#29 @prabhuramachandran
Hi @deniederhut -- I think I may have fixed the freezes, it was hard to reproduce but I would appreciate it if you could please try with vixen master to see if this does indeed fix the issue.  Unfortunately, you will need to test with Python 3.6 for now until I get #27 fixed with a Jigna release.
