### 2019-08-12 prabhuramachandran/vixen#29 @deniederhut
Looks like the file chooser still freezes for me. Is there some way to turn on more verbose logging so I have something more useful to send you?
