### 2019-08-12 prabhuramachandran/vixen#29 @prabhuramachandran
Would it be possible to test with the latest binary release?  https://github.com/vixen-project/vixen/releases  That may eliminate a build issue.

As far as logging goes, there isn't much going on to really be able to log.  I still do not have any idea of how you installed vixen, what Python version are you using (what distribution), can you also try the following in a Python shell:
```
import tkinter
print(tkinter.TkVersion)
print(tkinter.TclVersion)
```
Then you can try this:
```
import vixen.ui_utils as UU
UU.askdirectory()
```
Does that also freeze?

Under the covers, this is effectively what is called when you press "browse" for a directory.  Similarly for a file etc.  I wish there were a way of doing this from the browser but last I checked that is not easy.

Also, you should be able to edit the textbox left of the browse button and type out a directory -- just to see if the app is working.
