### 2019-08-13 prabhuramachandran/vixen#29 @deniederhut
Python 3.6.8, and I am installing from master.

Here are the tkinter versions:

```
>>> print(tkinter.TkVersion)
8.6
>>> print(tkinter.TclVersion)
8.6
```

Calling askdirectory from Python and the clicking cancel returns an empty string:

```
>>> import vixen.ui_utils as UU
>>> UU.askdirectory()
''"
```

and then the UI freezes. Calling it and choosing a directory returns the directory, raises a warning, and the UI freezes.

```
>>> import vixen.ui_utils as UU
>>> UU.askdirectory()
objc[1996]: Class FIFinderSyncExtensionHost is implemented in both /System/Library/PrivateFrameworks/FinderKit.framework/Versions/A/FinderKit (0x7fff9a321c90) and /System/Library/PrivateFrameworks/FileProvider.framework/OverrideBundles/FinderSyncCollaborationFileProviderOverride.bundle/Contents/MacOS/FinderSyncCollaborationFileProviderOverride (0x119e20cd8). One of the two will be used. Which one is undefined.
'/Users/dillon/test-images'
```
