### 2019-08-13 prabhuramachandran/vixen#29 @deniederhut
The macos binary looks like it is working -- are you vendoring some dependencies inside the app?
