### 2019-08-13 prabhuramachandran/vixen#29 @prabhuramachandran
@deniederhut -- Thanks so much for testing this and the information.  What distribution is this with? brew? or conda?  Also what version of MacOS?  I am still on High Sierra (10.13.6).

I am glad the binary is working, yes it vendorizes everything including Python, Tcl/Tk etc. (and fits in 25MB or less).  Honestly, our expectation is for people to use the binary as we do not expect people to be Python users.  In any case, your issue suggests that I should look around for a cleaner way to do this and also see how best to reproduce this issue.  I see there is a tkfilebrowser but was hoping to avoid any additional dependencies.
