### 2019-08-14 prabhuramachandran/vixen#29 @deniederhut
From anaconda - `Python 3.6.8 :: Anaconda, Inc`

Also High Sierra 10.13.6 but I really need to just bite the bullet and run the update.
