### 2019-08-22 #1501 @prabhuramachandran
Hi @xuanxu -- the syntax error is something I will fix by setting a python_requires this weekend as fixing the dependent package to support 3.7 opens a can of worms that requires much more time for me to address right now (the issue isn't vixen but the jigna package which doesn't support Python 3.7).

Issue 29 is something I am having a hard time reproducing so will try and spend some time on that but I have an alternative that does work, which is to use the binaries.  If it is OK to complete the reviews with the binaries, that may be more efficient.
