### 2019-08-22 #1501 @xuanxu
@prabhuramachandran great, as long as all that is documented, i.e. readme states clearly the required Python, the binaries installation steps, etc, I think is ok.
