### 2019-09-11 565c976
Set python_requires to temporarily fix #27.
