### 2019-09-11 prabhuramachandran/vixen#27 @prabhuramachandran
I just pushed a simple temporary fix for this until I get time to fix jigna.
