### 2019-09-11 #1501 @prabhuramachandran
I've just merge a tiny fix for that.  I think you should be able to test the software itself with the binaries available.  I will resolve the other issues as time permits.
