### 2019-09-16 #1501 @xuanxu
Ok, everything looks good @prabhuramachandran, here are the next steps:
- Please make a new tagged release that includes all changes made during this review
- Archive that release in Zenodo
- Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also add your ORCID.

Once you do that please report here the version number and the Zenodo DOI.
