### 2019-09-29 1bf0a49
Update changelog and version for 1.0-rc3 release.
