+++ vixen/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.0rc3.dev0'
+__version__ = '1.0rc3'
