### 2019-09-29 #1501 @prabhuramachandran
@xuanxu -- Thanks for the reminder.  I have uploaded version 1.0rc3 on pypi and also on zenodo here: http://doi.org/10.5281/zenodo.3464327 .  This is my first zenodo upload so I hope I did not mess anything up.  I will now make the binaries for people to use and upload them on the vixen github releases.
