+++ vixen/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.0rc3'
+__version__ = '1.0rc4.dev0'
