### 2019-09-30 #1501 @xuanxu
All ready for publication.
Pinging @openjournals/joss-eics for final acceptance.
