### 2019-10-01 #1501 @kyleniemeyer
Hi @prabhuramachandran, while going through the final checks I noticed the (sole) citation to a prior paper from last year, also about ViXeN.

How is this submission reporting on different scholarship than the article ["ViXeN: An open‐source package for managing multimedia data"](https://doi.org/10.1111/2041-210X.12892), published in _Methods in Ecology and Evolution_ ? They both appear to be software papers about the same software. While we do encourage authors to submit a software submission to JOSS and research articles using the software to other journals, the 2018 article seems to be just about the software.
