### 2019-10-01 #1501 @prabhuramachandran
@kyleniemeyer -- Thanks for checking!  The Methods in Ecology and Evolution paper was really about the application in the domain of ecology and camera trap videos/images and looked at a case study from the perspective of field biologists.  It talks about other tools relevant to the area and how this tool is useful in that context. The JOSS paper doesn't focus on the ecological application but on the fact that this is a general purpose multimedia annotation tool which is broader in scope.  We were given to understand that the primary purpose in JOSS is to review the software and not as an application in a particular domain of interest.  I am not sure if this answers the question.  Am happy to provide any clarifications.  Thanks.
