### 2019-10-02 #1501 @kyleniemeyer
@prabhuramachandran You're right that our purpose here is to review the software only, and not results—in fact, we do not really allow inclusion of results in our article (only examples). However, the concern here is that the other paper already describes the software in detail, in effect providing a "paper credit" for the software development, and a JOSS article would repeat that.

Has the scope or functionality of the software changed since it was described in the prior paper?
