### 2019-10-02 #1501 @prabhuramachandran
@kyleniemeyer -- I am sorry if I completely misunderstood the scope of JOSS papers and wasted everyone's time.  While we discussed the software in the previous paper, the review process did not go into  checking the software the way JOSS does or concern itself with archiving it etc. When we saw the nice talks at SciPy on JOSS, we thought we might submit it here to check it out.  At the time,  I recall reading an FAQ that said that it was OK to submit the software here despite there being another paper that discusses the software in a more traditional context. I assumed that given the different focus of the two, it was acceptable.  I am terribly sorry if I misunderstood.  It was totally not my intention to waste everyone's time (or get extra credit).  I leave it to you to decide about this submission as you see fit and will defer to your decision on this.  Apologies again.
