### 2019-10-02 #1501 @kyleniemeyer
@prabhuramachandran totally understand—we don't have entirely clear guidelines on this, as it hasn't come up too often. We'll discuss internally and get back to you. Either way, it looks like the software is great and the review process has made it even better :)
