### 2019-10-03 #1501 @kyleniemeyer
Hi @prabhuramachandran, the Associate Editors in Chief discussed this issue, and unfortunately we decided that we cannot publish this article, because we consider the prior article a software paper analogous to—albeit longer than—the JOSS article (even if the review process did not focus on the software like ours).

We do completely support JOSS authors in submitting research articles that use their published software, where the separate articles are about the results using the software and not the software itself. But that doesn't appear to be the case here.

I'm sorry for the confusion about this—we will work to improve our documentation to clarify our policy in such cases.

Regardless, we appreciate the efforts of @deniederhut and @amoeba in giving feedback on the software, and @xuanxu in handling this.
